[Refactor] Command에서 직접 진행하도록 수정
