+++ src/main/java/chess/domain/chessgame/ChessGame.java
@@ -1,16 +1,13 @@
 package chess.domain.chessgame;
 
-import chess.domain.Result;
-import chess.domain.Status;
 import chess.domain.chessboard.ChessBoard;
-import chess.domain.position.Position;
 import chess.factory.BoardFactory;
 import chess.view.InputView;
 import chess.view.OutputView;
 
 public class ChessGame {
 	private final ChessBoard chessBoard;
-	private ChessMenu chessMenu;
+	private Command command;
 
 	public ChessGame() {
 		this.chessBoard = BoardFactory.createBoard();
@@ -19,10 +16,10 @@ public class ChessGame {
 	public void play() {
 		OutputView.printRule();
 		initMenu();
-		chessMenu.validateStart();
+		command.validateStart();
 		OutputView.printBoard(chessBoard);
 
-		while (chessMenu.isNotEnd() && chessBoard.isLiveBothKing()) {
+		while (command.isNotEnd() && chessBoard.isLiveBothKing()) {
 			initMenu();
 			playRound();
 		}
@@ -35,7 +32,7 @@ public class ChessGame {
 
 	private boolean isNotAllowedMenu() {
 		try {
-			chessMenu = new ChessMenu(InputView.inputMenu());
+			command = new Command(InputView.inputMenu());
 			return false;
 		} catch (IllegalArgumentException | UnsupportedOperationException e) {
 			OutputView.printErrorMessage(e);
@@ -45,24 +42,10 @@ public class ChessGame {
 
 	private void playRound() {
 		try {
-			proceed();
+			command.execute(chessBoard);
 			OutputView.printBoard(chessBoard);
 		} catch (Exception e) {
 			OutputView.printErrorMessage(e);
-		}
-	}
-
-	private void proceed() {
-		if (chessMenu.isMove()) {
-			Position startPosition = chessMenu.createStartPosition();
-			Position targetPosition = chessMenu.createTargetPosition();
-			chessBoard.move(startPosition, targetPosition);
-		}
-
-		if (chessMenu.isStatus()) {
-			Status status = chessBoard.createStatus();
-			Result result = status.getResult();
-			OutputView.printResult(result);
 		}
 	}
 }
